Move enumeration_index into separate file
