+++ libvast/CMakeLists.txt
@@ -119,6 +119,7 @@ set(libvast_sources
     src/http.cpp
     src/icmp.cpp
     src/ids.cpp
+    src/index/enumeration_index.cpp
     src/io/read.cpp
     src/io/save.cpp
     src/io/write.cpp
@@ -338,6 +339,7 @@ set(tests
     test/http.cpp
     test/ids.cpp
     test/index/arithmetic_index.cpp
+    test/index/enumeration_index.cpp
     test/iterator.cpp
     test/json.cpp
     test/meta_index.cpp
