+++ libvast/src/value_index.cpp
@@ -256,50 +256,6 @@ string_index::lookup_impl(relational_operator op, data_view x) const {
     x);
 }
 
-// -- enumeration_index --------------------------------------------------------
-
-enumeration_index::enumeration_index(vast::type t, caf::settings opts)
-  : value_index{std::move(t), std::move(opts)},
-    index_{std::numeric_limits<enumeration>::max() + 1} {
-  // nop
-}
-
-caf::error enumeration_index::serialize(caf::serializer& sink) const {
-  return caf::error::eval([&] { return value_index::serialize(sink); },
-                          [&] { return sink(index_); });
-}
-
-caf::error enumeration_index::deserialize(caf::deserializer& source) {
-  return caf::error::eval([&] { return value_index::deserialize(source); },
-                          [&] { return source(index_); });
-}
-
-bool enumeration_index::append_impl(data_view x, id pos) {
-  if (auto e = caf::get_if<view<enumeration>>(&x)) {
-    index_.skip(pos - index_.size());
-    index_.append(*e);
-    return true;
-  }
-  return false;
-}
-
-caf::expected<ids>
-enumeration_index::lookup_impl(relational_operator op, data_view d) const {
-  return caf::visit(
-    detail::overload{
-      [&](auto x) -> caf::expected<ids> {
-        return make_error(ec::type_clash, materialize(x));
-      },
-      [&](view<enumeration> x) -> caf::expected<ids> {
-        if (op == in || op == not_in)
-          return make_error(ec::unsupported_operator, op);
-        return index_.lookup(op, x);
-      },
-      [&](view<list> xs) { return detail::container_lookup(*this, op, xs); },
-    },
-    d);
-}
-
 // -- address_index ------------------------------------------------------------
 
 address_index::address_index(vast::type t, caf::settings opts)
