+++ libvast/src/value_index_factory.cpp
@@ -20,6 +20,7 @@
 #include "vast/detail/type_traits.hpp"
 #include "vast/hash_index.hpp"
 #include "vast/index/arithmetic_index.hpp"
+#include "vast/index/enumeration_index.hpp"
 #include "vast/logger.hpp"
 #include "vast/type.hpp"
 #include "vast/value_index.hpp"
